Basic scaffold, no styling
Created the necessary collections, example material, and basic html layout for
the website. No styling has been added - this was purely to get content on the
page to make styling easier to add as we go.

Note: A lot of this filler content will be changed to be more useful/demonstrative
for the theme release so that users will be able to see the theme in full
action.